requirements.txt
Flask==1.1.2
numpy==1.19.0
opencv-python==4.5.1.48
pytesseract==0.3.7
python-dotenv==0.17.0
requests==2.25.1
requests-futures==1.0.0
Werkzeug==1.0.1